Provide separate code modules for Timers
