+++ .gitignore
@@ -1,6 +1,5 @@
 #Build files:
-main.o
-UARTComms.o
+*.o
 EEG-Simulator.elf
 EEG-Simulator.hex
 EEG-Simulator.map
